Extract KeyPair and PublicKey
